saving progress on logout
