+++ Backend/ManagerLayer/ManagerLayer.csproj
@@ -51,10 +51,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Login_Logout\UserLogoutManager.cs" />
+    <Compile Include="Models\LogoutManagerPOCOS.cs" />
     <Compile Include="Models\LoginManagerPOCOS.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="AccessControl\AuthorizationManager.cs" />
-    <Compile Include="Login\UserLoginManager.cs" />
+    <Compile Include="Login_Logout\UserLoginManager.cs" />
     <Compile Include="UserManagement\UserManagementManager.cs" />
     <Compile Include="Registration\UserRegistrationManager.cs" />
     <Compile Include="ValidationManager.cs" />
