+++ Backend/ManagerLayer/Models/LogoutManagerPOCOS.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ManagerLayer.Models
+{
+    class LogoutManagerPOCOS
+    {
+        
+
+    }
+
+    public class LogoutManagerResponsePOCO
+    {
+        public Guid id { get; set; }
+        public string message { get; set; }
+        public DateTime Timestamp { get; set; }
+
+
+    }
+}
